requirements.txt
aiogram==3.1.1
dnspython==2.4.2
pymongo==4.5.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==7.4.3